updates with JSON db functions

+++ Scripts/Update-SpeedyMigrations.ps1
@@ -8,7 +8,7 @@ $watch = [System.Diagnostics.Stopwatch]::StartNew()
 $scriptPath = Split-Path(Get-Variable MyInvocation).Value.MyCommand.Path
 $scriptPath = (Get-Item $scriptPath).Parent.FullName
 
-#$scriptPath = "C:\Workspaces\EpicCoders\Speedy"
+# $scriptPath = "C:\Workspaces\GitHub\Speedy"
 
 Write-Host $scriptPath
 
@@ -22,8 +22,20 @@ Write-Host $scriptPath
 $projectPath = "$scriptPath\Speedy.Client.Data.Sqlite"
 $startupPath = "$scriptPath\Speedy.IntegrationTests"
 
+if (!(Test-Path $projectPath))
+{
+	throw "$projectPath does not exists"	
+}
+
+if (!(Test-Path $startupPath))
+{
+	throw "$startupPath does not exists"	
+}
+
+# ii $projectPath
+
 Remove-Item "$projectPath\Migrations" -ErrorAction Ignore
-dotnet ef migrations add InitialMigration --framework "net6.0-windows" --configuration "debug" -c "ContosoClientDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
+dotnet ef migrations add InitialMigration --framework "net7.0-windows" --configuration "debug" -c "ContosoClientDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
 
 #
 # Client Data Sqlite Old
@@ -32,8 +44,18 @@ dotnet ef migrations add InitialMigration --framework "net6.0-windows" --configu
 $projectPath = "$scriptPath\Speedy.Client.Data.Sqlite.Old"
 $startupPath = "$scriptPath\Speedy.IntegrationTests"
 
+if (!(Test-Path $projectPath))
+{
+	throw "$projectPath does not exists"	
+}
+
+if (!(Test-Path $startupPath))
+{
+	throw "$startupPath does not exists"	
+}
+
 Remove-Item "$projectPath\Migrations" -ErrorAction Ignore
-dotnet ef migrations add InitialMigration --framework "net5.0-windows" --configuration "debug" -c "ContosoClientDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
+dotnet ef migrations add InitialMigration --framework "net48" --configuration "debug" -c "ContosoClientDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
 
 
 #
@@ -43,8 +65,18 @@ dotnet ef migrations add InitialMigration --framework "net5.0-windows" --configu
 $projectPath = "$scriptPath\Speedy.Website.Data.Sql"
 $startupPath = "$scriptPath\Speedy.Website"
 
+if (!(Test-Path $projectPath))
+{
+	throw "$projectPath does not exists"	
+}
+
+if (!(Test-Path $startupPath))
+{
+	throw "$startupPath does not exists"	
+}
+
 Remove-Item "$projectPath\Migrations" -ErrorAction Ignore
-dotnet ef migrations add InitialMigration --framework "net6.0" --configuration "debug" -c "ContosoSqlDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
+dotnet ef migrations add InitialMigration --framework "net7.0" --configuration "debug" -c "ContosoSqlDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
 
 #
 # Website Data Sql Old
@@ -53,8 +85,18 @@ dotnet ef migrations add InitialMigration --framework "net6.0" --configuration "
 $projectPath = "$scriptPath\Speedy.Website.Data.Sql.Old"
 $startupPath = "$scriptPath\Speedy.Website"
 
+if (!(Test-Path $projectPath))
+{
+	throw "$projectPath does not exists"	
+}
+
+if (!(Test-Path $startupPath))
+{
+	throw "$startupPath does not exists"	
+}
+
 Remove-Item "$projectPath\Migrations" -ErrorAction Ignore
-dotnet ef migrations add InitialMigration --framework "net5.0" --configuration "debug" -c "ContosoSqlDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
+dotnet ef migrations add InitialMigration --framework "netcoreapp3.1" --configuration "debug" -c "ContosoSqlDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
 
 #
 # Website Data Sqlite
@@ -63,8 +105,18 @@ dotnet ef migrations add InitialMigration --framework "net5.0" --configuration "
 $projectPath = "$scriptPath\Speedy.Website.Data.Sqlite"
 $startupPath = "$scriptPath\Speedy.Website"
 
+if (!(Test-Path $projectPath))
+{
+	throw "$projectPath does not exists"	
+}
+
+if (!(Test-Path $startupPath))
+{
+	throw "$startupPath does not exists"	
+}
+
 Remove-Item "$projectPath\Migrations" -ErrorAction Ignore
-dotnet ef migrations add InitialMigration --framework "net6.0" --configuration "debug" -c "ContosoSqliteDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
+dotnet ef migrations add InitialMigration --framework "net7.0" --configuration "debug" -c "ContosoSqliteDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
 
 #
 # Website Data Sqlite Old
@@ -73,5 +125,15 @@ dotnet ef migrations add InitialMigration --framework "net6.0" --configuration "
 $projectPath = "$scriptPath\Speedy.Website.Data.Sqlite.Old"
 $startupPath = "$scriptPath\Speedy.Website"
 
+if (!(Test-Path $projectPath))
+{
+	throw "$projectPath does not exists"	
+}
+
+if (!(Test-Path $startupPath))
+{
+	throw "$startupPath does not exists"	
+}
+
 Remove-Item "$projectPath\Migrations" -ErrorAction Ignore
-dotnet ef migrations add InitialMigration --framework "net5.0" --configuration "debug" -c "ContosoSqliteDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
+dotnet ef migrations add InitialMigration --framework "netcoreapp3.1" --configuration "debug" -c "ContosoSqliteDatabase" -p $projectPath -s $startupPath -o "$projectPath\Migrations"
